refactor: move affineTransform files to align folder
