+++ src/align/affineTransfrom/__tests__/__image_snapshots__/get-affine-transform-test-ts-debug-true-1-snap.png
