+++ src/align/affineTransfrom/__tests__/__image_snapshots__/get-affine-transform-test-ts-rgb-images-1-snap.png
