+++ src/align/affineTransfrom/__tests__/__image_snapshots__/get-affine-transform-test-ts-various-polygons-rotation-10-degrees-1-snap.png
