+++ src/align/affineTransfrom/__tests__/__image_snapshots__/get-affine-transform-test-ts-various-polygons-twice-same-image-1-snap.png
