+++ src/align/affineTransfrom/__tests__/__image_snapshots__/ransac-test-ts-2-d-data-points-polygon-rotated-180-degrees-1-snap.png
