+++ src/align/affineTransfrom/getEuclideanDistance.ts
